swav_offline_pipeline/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.4.0+cu121
torchvision==0.19.0+cu121
torchaudio==2.4.0+cu121
xformers==0.0.27.post2
diffusers
transformers
accelerate
safetensors
pillow
numpy
opencv-python
pyyaml
rich
moviepy
ffmpeg-python
pydub
librosa
soundfile
TTS==0.22.0
huggingface_hub
realesrgan
gdown
pandas
python-slugify
